Intent: centralize logging configuration\nOperators involved: [Self-organization]\nAffected invariants: [#8]\nKey changes:\n- Extracted _configure_root into logging_base.py\n- Updated logging.py and logging_utils.py to import shared configuration\n- Adjusted tests to reload shared module\nExpected risks/dissonances: Shared state across modules; mitigated by reloading base in tests\nMetrics: Stable C(t), deterministic νf\nEquivalence map: _configure_root -> tnfr.logging_base._configure_root
